--- src/main/java/org/prgms/voucherProgram/domain/CustomerMenuType.java
@@ -1,4 +1,0 @@
-package org.prgms.voucherProgram.domain;
-
-public enum CustomerMenuType {
-}
