+++ src/main/java/org/prgms/voucherProgram/domain/menu/CustomerMenuType.java
@@ -1,0 +1,4 @@
+package org.prgms.voucherProgram.domain.menu;
+
+public enum CustomerMenuType {
+}
